Move OverlayCamera into Cameras module
