+++ include/BLIB/Cameras.hpp
@@ -17,5 +17,6 @@
 #include <BLIB/Cameras/3D/CameraController3D.hpp>
 #include <BLIB/Cameras/3D/Controllers/OrbiterController.hpp>
 #include <BLIB/Cameras/Camera.hpp>
+#include <BLIB/Cameras/OverlayCamera.hpp>
 
 #endif
