+++ include/BLIB/Cameras/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(BLIB PUBLIC
     Camera.hpp
+    OverlayCamera.hpp
 )
 
 add_subdirectory(2D)
