+++ include/BLIB/Render/Overlays/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(BLIB PUBLIC
 	Overlay.hpp
-	OverlayCamera.hpp
 	OverlayObject.hpp
 )
