+++ src/Cameras/CMakeLists.txt
@@ -1,5 +1,6 @@
 target_sources(BLIB PRIVATE
     Camera.cpp
+    OverlayCamera.cpp
 )
 
 add_subdirectory(2D)
