+++ src/Render/Overlays/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(BLIB PRIVATE
 	Overlay.cpp
-	OverlayCamera.cpp
 	OverlayObject.cpp
 )
